refactor: break apart into two linters
