+++ README.md
@@ -105,7 +105,8 @@ Other dedicated linters that are built-in are:
 | [biomejs][biomejs]                     | `biomejs`              |
 | [blocklint][blocklint]                 | `blocklint`            |
 | [buf_lint][buf_lint]                   | `buf_lint`             |
-| [buf_api_linter][buf_api_linter]                   | `buf_api_linter`             |
+| [api_linter][api_linter]               | `api_linter`           |
+| [api_linter_buf][api_linter_buf]       | `api_linter_buf`       |
 | [buildifier][buildifier]               | `buildifier`           |
 | [cfn-lint][cfn-lint]                   | `cfn_lint`             |
 | [cfn_nag][cfn_nag]                     | `cfn_nag`              |
@@ -607,7 +608,8 @@ vimcats -t -f lua/lint.lua lua/lint/parser.lua > doc/lint.txt
 [verilator]: https://verilator.org/guide/latest/
 [actionlint]: https://github.com/rhysd/actionlint
 [buf_lint]: https://github.com/bufbuild/buf
-[buf_api_linter]: https://github.com/googleapis/api-linter
+[api_linter]: https://github.com/googleapis/api-linter
+[api_linter_buf]: https://github.com/googleapis/api-linter
 [erb-lint]: https://github.com/shopify/erb-lint
 [tfsec]: https://github.com/aquasecurity/tfsec
 [tlint]: https://github.com/tighten/tlint
